requirements.txt
asgiref==3.4.1
autopep8==1.6.0
backports.zoneinfo==0.2.1
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
face-recognition==1.3.0
face-recognition-models==0.3.0
openpyxl==3.0.5
psycopg2==2.9.1
django-heroku==0.3.1
whitenoise==5.2.0
setuptools==56.0.0